fix_locmax_error
