+++ lib/widgets/weather/weather_chart_empty_state.dart
@@ -1,0 +1,50 @@
+/// Weather chart empty state widget.
+///
+// Time-stamp: <Friday 2026-01-24 10:00:00 +1100>
+///
+/// Copyright (C) 2026, Software Innovation Institute, ANU.
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License").
+///
+/// License: https://opensource.org/license/gpl-3-0.
+
+library;
+
+import 'package:flutter/material.dart';
+
+/// Widget displayed when there is no data available for the chart.
+class WeatherChartEmptyState extends StatelessWidget {
+  const WeatherChartEmptyState({required this.dataTitle, super.key});
+
+  final String dataTitle;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Padding(
+        padding: const EdgeInsets.all(24),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Icon(Icons.info_outline, size: 48, color: Colors.grey[400]),
+            const SizedBox(height: 16),
+            Text(
+              'No ${dataTitle.toLowerCase()} data available',
+              style: Theme.of(
+                context,
+              ).textTheme.titleMedium?.copyWith(color: Colors.grey[600]),
+            ),
+            const SizedBox(height: 8),
+            Text(
+              'This might be because the data is not available for this time period.',
+              textAlign: TextAlign.center,
+              style: Theme.of(
+                context,
+              ).textTheme.bodySmall?.copyWith(color: Colors.grey[500]),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
