+++ lib/widgets/weather/weather_chart_header.dart
@@ -1,0 +1,52 @@
+/// Weather chart header widget.
+///
+// Time-stamp: <Friday 2026-01-24 10:00:00 +1100>
+///
+/// Copyright (C) 2026, Software Innovation Institute, ANU.
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License").
+///
+/// License: https://opensource.org/license/gpl-3-0.
+
+library;
+
+import 'package:flutter/material.dart';
+
+import 'package:intl/intl.dart';
+
+/// Header widget for the weather chart showing title and date range.
+class WeatherChartHeader extends StatelessWidget {
+  const WeatherChartHeader({
+    required this.title,
+    required this.startDate,
+    required this.endDate,
+    super.key,
+  });
+
+  final String title;
+  final DateTime startDate;
+  final DateTime endDate;
+
+  @override
+  Widget build(BuildContext context) {
+    final dateFormat = DateFormat('MMM dd');
+
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        Text(
+          title,
+          style: Theme.of(
+            context,
+          ).textTheme.titleMedium?.copyWith(fontWeight: FontWeight.bold),
+        ),
+        Text(
+          '${dateFormat.format(startDate)} - ${dateFormat.format(endDate)}',
+          style: Theme.of(
+            context,
+          ).textTheme.bodySmall?.copyWith(color: Colors.grey[600]),
+        ),
+      ],
+    );
+  }
+}
